digital-twin-platform/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
pymongo==4.1.1
paho-mqtt==1.6.1
numpy==1.22.3
pandas==1.4.2
eventlet==0.33.0
python-engineio==4.3.1
python-socketio==5.5.2
requests==2.27.1 